requirements.txt
Click
click
numpy
pyyaml
pytest
matplotlib
seaborn
notebook
torch
pytorch-lightning
tensorboard
neptune
tqdm
scipy
